+++ packages/caelus/src/index.ts
@@ -19,3 +19,4 @@ export * from "./lots.js";
 export * from "./profections.js";
 export * from "./firdaria.js";
 export * from "./releasing.js";
+export * from "./vedic.js";